feat : auth navigator added
